refactor: remove unused contexts, update dropdown
