+++ src/components/custom/Layout/Header.tsx
@@ -11,11 +11,11 @@ import Wallet from '../Wallet';
 import config from '../../../../config';
 
 import ConnectButton from '../ConnectButton';
-import { useUser } from '../../../shared/@ocean/context/UserContext';
+import { useWeb3 } from '../../../shared/@ocean/context/WalletContext';
 
 const Navigation = () => {
   const Profile = config.routes.profile;
-  const { user } = useUser();
+  const { user } = useWeb3();
   return (
     <div className="d-flex flex-column flex-md-row align-center">
       <div>{!user ? <ConnectButton /> : <Wallet />}</div>
