+++ src/components/custom/ReportSection/Dropdown/styles.module.scss
@@ -1,0 +1,3 @@
+.dropdown {
+  max-width: 100%;
+}
