+++ src/components/custom/Wallet/Account.tsx
@@ -4,11 +4,11 @@ import Caret from '../../../assets/caret.svg';
 import CaretBlack from '../../../assets/caret_black.svg';
 import WalletIcon from '../../../assets/walletIcon.svg';
 import styles from './Account.module.scss';
-import { useAccount, useEnsName, useEnsAvatar } from 'wagmi';
+import { useEnsName, useEnsAvatar } from 'wagmi';
 import Image from 'next/image';
 import Avatar from '../Avatar';
 import { truncateWalletAddress } from '../../../shared/utilities/truncateAddress';
-import { useUser } from '../../../shared/@ocean/context/UserContext';
+import { useWeb3 } from '../../../shared/@ocean/context/WalletContext';
 
 type AccountPropType = {
   mobile?: boolean;
@@ -17,8 +17,7 @@ type AccountPropType = {
 // Forward ref for Tippy.js
 // eslint-disable-next-line
 export default function Account({ mobile }: AccountPropType): ReactElement {
-  const { address: accountId } = useAccount();
-  const { user } = useUser();
+  const { user } = useWeb3();
   const { data: accountEns } = useEnsName({
     address: user as `0x${string}`,
     chainId: 1,
