+++ src/shared/@ocean/context/WalletContext.tsx
@@ -10,7 +10,6 @@ import { magic } from '../../utilities/libs/magic';
 import { WalletInfo } from 'magic-sdk';
 import { Signer, ethers } from 'ethers';
 
-// Define the structure of the Web3 context state
 type Web3ContextType = {
   web3: Web3 | null;
   initializeWeb3: () => void;
@@ -21,31 +20,27 @@ type Web3ContextType = {
   chainId: number;
   web3Signer: Signer;
   isWalletConnecting: boolean;
+  user: string;
 };
 
-// Create the context with default values
 const Web3Context = createContext<Web3ContextType>({} as Web3ContextType);
 
-// Custom hook to use the Web3 context
 export const useWeb3 = () => useContext(Web3Context);
 
-// Provider component to wrap around components that need access to the context
 export const Web3Provider = ({ children }: { children: React.ReactNode }) => {
-  // State variable to hold an instance of Web3
   const [web3, setWeb3] = useState<Web3 | null>(null);
   const [chainId, setChainId] = useState<number>(null);
   const [isWalletConnecting, setIsWalletConnecting] = useState(false);
+  const [user, setUser] = useState<string | null>(null);
   const [web3Signer, setWeb3Signer] = useState<Signer>();
   const [walletConnectionType, setWalletConnectionType] =
     useState<WalletInfo>(undefined);
 
-  // Initialize Web3
   const initializeWeb3 = useCallback(async () => {
-    // Get the provider from the Magic instance
     const provider = await magic.wallet.getProvider();
     const ethersMagicProvider = new ethers.providers.Web3Provider(provider);
     const magicEthersSigner = ethersMagicProvider.getSigner();
-    // Create a new instance of Web3 with the provider
+
     const web3 = new Web3(provider);
 
     const isLoggedIn = await magic.user.isLoggedIn();
@@ -56,7 +51,6 @@ export const Web3Provider = ({ children }: { children: React.ReactNode }) => {
       setWalletConnectionType(walletInfo);
       setWeb3Signer(magicEthersSigner);
     }
-    // Save the instance to state
     setWeb3(web3);
   }, []);
 
@@ -89,15 +83,25 @@ export const Web3Provider = ({ children }: { children: React.ReactNode }) => {
     }
   };
 
-  // Effect to initialize Web3 when the component mounts
+  const fetchUserAccount = async () => {
+    const accounts = await web3?.eth.getAccounts();
+
+    setUser(accounts ? accounts[0] : null);
+  };
+
   useEffect(() => {
     initializeWeb3();
   }, []);
+
+  useEffect(() => {
+    fetchUserAccount();
+  }, [web3]);
 
   return (
     <Web3Context.Provider
       value={{
         web3,
+        user,
         initializeWeb3,
         handleConnect,
         handleDisconnect,
